Updated scripts

+++ post-installation-debian.sh
@@ -30,7 +30,11 @@ ask_yes_no () {
 #
 if ask_yes_no "Change hostname?"; then
 	read -p "Enter new hostname: " new_hostname
-	sudo hostnamectl hostname $new_hostname
+	if [ -f "/etc/hostname" ] && pacman -Q | grep -qF systemd; then
+		sudo hostnamectl hostname $new_hostname
+	else
+		echo "Hostname file doesn't exist. Something went wrong."
+	fi
 else
 	echo "Skipping."
 fi
@@ -38,10 +42,15 @@ fi
 # Ask the user if they'd like to install docker. First we update the system, get the keyring for docker, add the
 # repository, and install the packages through apt.
 #
+if ask_yes_no "Update & upgrade system?"; then
+	sudo apt-get update && sudo apt-get upgrade -y
+else
+	echo "Skipping."
+fi
+
 if ask_yes_no "Install docker?"; then
 	echo "Installing docker engine"
 
-	sudo apt-get update
 	sudo apt-get install ca-certificates curl
 	sudo install -m 0755 -d /etc/apt/keyrings
 	sudo curl -fsSL https://download.docker.com/linux/debian/gpg -o /etc/apt/keyrings/docker.asc
@@ -69,46 +78,73 @@ fi
 # User chosen or default basic packages.
 #
 if ask_yes_no "Install packages?"; then
-	if ask_yes_no "Install default packages?\n(ncdu, htop, ufw, qemu-guest-agent, nvim, vim, fzf)?"; then
+	if ask_yes_no "Install default packages? (neovim htop neofetch ncdu qemu-guest-agent git openssh)?"; then
 		echo "Installing default pacakges."
-		sudo apt update && sudo apt upgrade -y
-		sudo apt install ncdu htop ufw qemu-guest-agent neovim vim fzf -y
-		echo "Enabling qemu-guest-agent."
-		sudo systemctl enable --now qemu-guest-agent.service
-	else
-		read -p "Type the packaged you would like installed (separated by a single space): " packages_to_install
-		echo "installing $packages_to_install."
-		sudo apt install $packages_to_isntall -y
-		if sudo apt list --installed | grep -qF "qemu-guest-agent"; then
-			sudo systemctl enable --now qemu-guest-agent.service
+		sudo apt install neovim htop neofetch ncdu qemu-guest-agent git openssh
+		if ask_yes_no "Install more packages?"; then
+			read -p "Type the packages you would like to install (separated by a single space): " packages_to_install
+			sudo pacman -S $packages_to_install
 		else
 			:
 		fi
+	else
+		read -p "Type the packages you would like to install (separated by a single space): " packages_to_install
+		sudo apt install $packages_to_install
 	fi
 else
-	echo "No packages chosen for install, updating & upgrading system."
-	sudo apt update && sudo apt upgrade -y
+	echo "No packages chosen for install. Skipping."
 fi
+#
+# Detect and ask if the user wants to enable these services in systemd
+#
+if apt list --installed | grep -qF openssh; then
+	if ask_yes_no "OpenSSH installation detected, enable now?"; then
+		sudo systemctl enable --now sshd
+	else
+		echo "Skipping."
+	fi
+fi
+
+if apt list --installed | grep -qF docker; then
+	if ask_yes_no "Docker installation detected, enable now?"; then
+		sudo systemctl enable --now docker.service
+	else
+		:
+	fi
+fi
+
+if apt list --installed | grep -qF neovim; then
+	if ask_yes_no "NeoVIM installation detected, do you want 'v' 'vi' and 'vim' to be aliases for the 'nvim' command?"; then
+		if grep "alias vi='nvim'" /home/$USER/.bashrc && grep "alias vim='nvim'" /home/$USER/.bashrc && grep "alias v='nvim'" /home/$USER/.bashrc; then
+			echo "Aliases already exist. Skipping."
+		else
+			echo "alias vi='nvim'" >> /home/$USER/.bashrc
+			echo "alias vim='nvim'" >> /home/$USER/.bashrc
+			echo "alias v='nvim'" >> /home/$USER/.bashrc
+		fi
+	else
+		echo "Skipping."
+	fi
+else
+	:
+fi
+
 # *TESTING*
 #
 # Ask to install the custom bash prompt. This is a "debian red" prompt.
 #
 if ask_yes_no "Add custom red bash prompt?"; then
 	echo "Adding custom red bash prompt."
-
-	if [ -f "/home/$USER/.bashrc" ]; then
+	if [ -f "/home/$USER/.bashrc" ] && grep -qF "PS1='/[\e" "/home/$USER/.bashrc"; then
 		cp /home/$USER/.bashrc /home/$USER/.bashrc.old
-		if grep -qF "#CUSTOM_BASH_PROMPT" "/home/$USER/.bashrc"; then
-			echo "bash prompt already exists"
-		else
-			echo "#CUSTOM_BASH_PROMPT" >> /home/$USER/.bashrc
-			echo "PS1='\[\e[91m\]\u@\h\[\e[0m\]:\[\e[38;5;38m\]\w\[\e[0m\]\$ '" >> /home/$USER/.bashrc
-		fi
+		awk '{if ($0 ~ /PS1/} print "PS1='\''\[\\e[91m\\]\\u@\\h\\[\\e[0m\\]:\\[\\e[38;5;38m\\]\\w\\[\\e[0m\\]\\$ '\''"; else print}' /home/$USER/.bashrc > tmpfile && mv tmpfile /home/$USER/.bashrc
+	elif [ -f "/home/$USER/.bashrc" ]; then
+		echo "PS1='\[\e[91m\]\u@\h\[\e[0m\]:\[\e[38;5;38m\]\w\[\e[0m\]\$ '" >> /home/$USER/.bashrc
 	else
 		echo ".bashrc does not exist. Something went wrong."
 	fi
 else
-	echo "Keeping default bash prompt."
+	echo "Skipping."
 fi
 #
 # Ask the user if they'd like to add an automatic session timeout in a user defined number of seconds.
@@ -119,7 +155,12 @@ if ask_yes_no "Set a console timeout?"; then
 		if grep -qF "TMOUT=$tmout_seconds" "/home/$USER/.bashrc"; then
 			echo "TMOUT=$tmout_seconds already exists in ~/.bashrc. Skipping."
 		elif grep -qF "TMOUT=" "/home/$USER/.bashrc"; then
-			echo "Some timeout (TMOUT=) line already exists in ~/.bashrc. Skipping."
+			if ask_yes_no "A timeout already exists in ~/.bashrc. Overwrite?"; then
+				sed -i "s/TMOUT.*/TMOUT=$tmout_seconds/" ~/.bashrc
+				echo "A $tmout_seconds second timeout has been added and will be applied on next login or by running a new shell."
+			else
+				echo "Skipping."
+			fi
 		else
 			echo "TMOUT=$tmout_seconds" >> /home/$USER/.bashrc
 			echo "A $tmout_seconds second timeout has been added and will be applied on next login or by running a new shell."
@@ -161,7 +202,7 @@ ssh_key_setup() {
 #
 # Backup any authorized_keys file that may exist and create a new one with user defined keys.
 # 
-if ask_yes_no "Modify SSH keys? This will create a backup of ~/.ssh/authorized_keys and create a new one."; then
+if ask_yes_no "Modify authorized SSH keys? This will create a backup of ~/.ssh/authorized_keys and create a new one."; then
 	echo "Creating ~/.ssh/authorized_keys"
 	if [ -f "/home/$USER/.ssh/authorized_keys" ]; then
 		mv /home/$USER/.ssh/authorized_keys /home/$USER/.ssh/authorized_keys.old
@@ -183,7 +224,6 @@ if ask_yes_no "Change SSH port?"; then
 	read -p "Port to use for SSH: " ssh_port
 	echo "Changing default SSH port to $ssh_port."
 	if [ -f "/etc/ssh/sshd_config" ]; then
-#		sudo sed -i 's/#Port 22/Port $ssh_port/' /etc/ssh/sshd_config
 		sudo sed -i "s/.*Port [0-9]*/Port $ssh_port/" /etc/ssh/sshd_config
 		echo "SSH changed to port $ssh_port"
 	else
@@ -273,5 +313,5 @@ else
 fi
 
 echo "****END OF SCRIPT****"
-
+bash
 exit
